bug fix:  limit queries to 3 scanned text elements
