+++ .env
@@ -1,1 +1,8 @@
 chromedriver_path=/usr/local/bin/chromedriver
+AWS_ACCESS_KEY_ID=AKIAY5ZLAZLVRN2VOMGO
+AWS_SECRET_ACCESS_KEY=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
+AWS_DEFAULT_REGION=us-east-2
+DS_DB_NAME=rxidDS
+DS_DB_USER=rxidDS
+DS_DB_HOST=rxidds.cqqygklpjkea.us-east-2.rds.amazonaws.com
+DS_DB_PASSWORD=rxid_lambda
