+++ .gitignore
@@ -122,3 +122,4 @@ Web Service Endpoint/00000001.jpg
 Web Service Endpoint/sample_pill_image.jpg
 Web Service Endpoint/gcvision.cred.json
 sample_pill_image.jpg
+.env
